wrap up connections
window resize resizes graph canvas
add popups
hit esc to close popups
parse dates
cleanup all the things
remove old connections file
